% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_fcts.R
\name{available_counties}
\alias{available_counties}
\title{Data of Available Countries - Should I Want to Expand the Tool}
\usage{
available_counties()
}
\value{
Tibble with countries and their currencies.
}
\description{
'country' must be of two characters to stay consistent with the code flow.
'locale' must be one of the options availablke here:
https://www.w3schools.com/jsref/jsref_tolocalestring_number.asp
}
