% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pl_settings_default.R
\name{calc_pl_deductions}
\alias{calc_pl_deductions}
\title{Calculate PL Deductions Breakdown Given Gross Annual Earnings}
\usage{
calc_pl_deductions(
  annual_earnings,
  alpha_scheme = TRUE,
  standard_tax = TRUE,
  settings = pl_settings
)
}
\arguments{
\item{annual_earnings}{Vector of annual gross earnings.}

\item{alpha_scheme}{Not in use for the PL deductions.}

\item{standard_tax}{Use progressive tax (UoP) or linear tax (B2B)?}

\item{settings}{Use either default setting for the PL or the data supplied by the user.}
}
\value{
Tibbles with deductions and net income values.
}
\description{
Note that the arguments must be the same across all \code{calc_\{country\}_deductions}
to ensure that all Shiny settings are reversible.
}
\examples{
\dontrun{
calc_pl_deductions(annual_earnings = c(NA, 70000, 80000, 90000, NA))$df_deductions
}
}
