% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_default_uk_settings.R
\name{calc_uk_deductions}
\alias{calc_uk_deductions}
\title{Calculate UK Deductions Breakdown Given Gross Annual Earnings}
\usage{
calc_uk_deductions(
  annual_earnings,
  alpha_scheme = TRUE,
  standard_tax = TRUE,
  incluse_slp2 = TRUE,
  incluse_slp3 = TRUE,
  settings_user = FALSE
)
}
\arguments{
\item{annual_earnings}{Vector of annual gross earnings.}

\item{alpha_scheme}{Should the pension contribution be deducted using alpha scheme?}

\item{standard_tax}{Not in use for the UK deductions.}

\item{incluse_slp2}{Include Student Loan Plan 2 in the breakdown?}

\item{incluse_slp3}{Include Student Loan Plan 3 in the breakdown?}

\item{settings_user}{Use either default setting for the UK or the ones supplied
by the user.}
}
\value{
Tibbles with deduction and net income values by either the category
or the category wide (see \code{uk_deduction_types}).
}
\description{
Note that the arguments must be the same across all \code{calc_\{country\}_deductions}
functions to ensure that all Shiny settings are reversible.
}
\examples{
\dontrun{
calc_uk_deductions(
   annual_earnings = c(30000, 35000, 50000),
   alpha_scheme    = TRUE,
   standard_tax    = TRUE,
   incluse_slp2    = TRUE,
   incluse_slp3    = TRUE,
   settings_user   = FALSE
)$df_deductions_category_wide
}
}
