% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_fcts.R
\name{custom_spinner}
\alias{custom_spinner}
\title{Define Custom Spinner Animation}
\usage{
custom_spinner(ui_element, col = palette_global$body_tertiary_bg)
}
\arguments{
\item{ui_element}{UI to attache the spinner to.}

\item{col}{Spinner colour.}
}
\description{
Define Custom Spinner Animation
}
