% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_fcts.R
\name{div_with_icon}
\alias{div_with_icon}
\title{Insert Div with Icon and Link}
\usage{
div_with_icon(..., link, icon_size = "1rem", tt_text = "Find out more!")
}
\arguments{
\item{...}{Any UI element.}

\item{link}{Link to the website.}

\item{icon_size}{Icon size.}

\item{tt_text}{Text in tooltip.}
}
\value{
Div.
}
\description{
Insert Div with Icon and Link
}
