% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_prep_df_main.R
\name{fit_earnings_dist}
\alias{fit_earnings_dist}
\title{Fit a Spline to Available Deciles to Find In-between Values}
\usage{
fit_earnings_dist(available_deciles)
}
\arguments{
\item{available_deciles}{Named vector/list with deciles. The names should be
of the form *decile*th, e.g., *10th*, *20th*, etc. See examples.}
}
\value{
Tibble with actual deciles and interpolated values.
}
\description{
Fit a Spline to Available Deciles to Find In-between Values
}
\examples{
\dontrun{
decile_vector <- c("10th" = 200, "20th" = 400, "50th" = 800)
fit_earnings_dist(decile_vector)
}
}
