% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prep_df_main.R
\name{get_df_earnings_dist}
\alias{get_df_earnings_dist}
\title{Generate Join Earnings Distribution & Deductions Data}
\usage{
get_df_earnings_dist(settings_from = uk_settings, settings_to = pl_settings)
}
\arguments{
\item{settings_from}{Country from settings.}

\item{settings_to}{Country to settings.}
}
\value{
Join earnings distribution tibble with deciles and deduction breakdown.
This is be the main data in the app that will rely on the user input and update
on the button click. This function will also return a helper legend-table.
}
\description{
Generate Join Earnings Distribution & Deductions Data
}
\examples{
\dontrun{
get_df_earnings_dist(uk_settings, pl_settings)$df_main
get_df_earnings_dist(uk_settings, pl_settings)$df_cat_table
}
}
