% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_utils_fcts.R
\name{picker_options_with_flags}
\alias{picker_options_with_flags}
\title{Add Flags to the pickerInput Options}
\usage{
picker_options_with_flags(flag_class, country)
}
\arguments{
\item{flag_class}{Class name, e.g. 'fi fi-pl'.}

\item{country}{Country name.}
}
\description{
Add Flags to the pickerInput Options
}
