% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pl_df_expend}
\alias{pl_df_expend}
\title{PL Expenditure Data}
\format{
Data frame. See \code{data-raw/prep_pl_df_expend.R} for detailed derivations.
\describe{
  \item{expenditure}{Expenditure category.}
  \item{percentile}{Percentile. Integer from 0 to 100.}
  \item{actual_values}{Actual published percentile values.}
  \item{interpolated_values}{Values interpolated using the spline method.}
  \item{avg}{Published average expenditure per adult.}
}
}
\usage{
pl_df_expend
}
\description{
PL Expenditure Data
}
\keyword{datasets}
