% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uk_df_expend}
\alias{uk_df_expend}
\title{UK Expenditure Data}
\format{
Data frame. See \code{data-raw/prep_uk_df_expend.R} for detailed derivations.
\describe{
  \item{expenditure}{Expenditure category.}
  \item{percentile}{Percentile. Integer from 0 to 100.}
  \item{actual_values}{Actual published percentile values.}
  \item{interpolated_values}{Values interpolated using the spline method.}
  \item{avg}{Published average expenditure per adult.}
  \item{scaling_factor}{Derived column indicating the proportion expenditure change in respect to the average.
  Used to scale up expenditure in countries which do not publish detailed values.}
}
}
\usage{
uk_df_expend
}
\description{
UK Expenditure Data
}
\keyword{datasets}
