% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_pl_settings_default.R
\name{calc_pl_deductions}
\alias{calc_pl_deductions}
\title{Calculate PL Deductions Breakdown Given Gross Annual Earnings}
\usage{
calc_pl_deductions(
  annual_earnings,
  alpha_scheme = TRUE,
  standard_tax = TRUE,
  incluse_slp2 = TRUE,
  incluse_slp3 = TRUE,
  user_data = NULL
)
}
\arguments{
\item{annual_earnings}{Vector of annual gross earnings.}

\item{alpha_scheme}{Not in use for the PL deductions.}

\item{standard_tax}{Use progressive tax (UoP) or linear tax (B2B)?}

\item{incluse_slp2}{Include Student Loan Plan 2 in the breakdown?}

\item{incluse_slp3}{Include Student Loan Plan 3 in the breakdown?}

\item{user_data}{Use either default setting for the UK (\code{NULL}) or the
data supplied by the user.}
}
\value{
Tibbles with deduction and net income values by either the category
or the category wide (see \code{pl_deduction_types}).
}
\description{
Note that the arguments must be the same across all \code{calc_\{country\}_deductions}
to ensure that all Shiny settings are reversible.
}
\examples{
\dontrun{
calc_pl_deductions(
   annual_earnings = c(70000, 80000, 90000),
   alpha_scheme    = TRUE,
   standard_tax    = TRUE,
   incluse_slp2    = TRUE,
   incluse_slp3    = TRUE,
   user_data       = NULL
)$df_deductions_category_wide
}
}
