% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_uk_settings_default.R
\name{calc_uk_deductions}
\alias{calc_uk_deductions}
\title{Calculate UK Deductions Breakdown Given Gross Annual Earnings}
\usage{
calc_uk_deductions(
  annual_earnings,
  alpha_scheme = TRUE,
  standard_tax = TRUE,
  settings = uk_settings
)
}
\arguments{
\item{annual_earnings}{Vector of annual gross earnings.}

\item{alpha_scheme}{Should the pension contribution be deducted using alpha scheme?}

\item{standard_tax}{Not in use for the UK deductions.}

\item{settings}{Use either default setting for the UK or the data supplied by the user.}
}
\value{
Tibbles with deductions and net income values.
}
\description{
Note that the arguments must be the same across all \code{calc_\{country\}_deductions}
functions to ensure that all Shiny settings are reversible.
}
\examples{
\dontrun{
calc_uk_deductions(annual_earnings = c(NA, 30000, 35000, 50000, NA))$df_deductions
}
}
