% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pl_settings}
\alias{pl_settings}
\title{Default PL Settings}
\format{
Nested named list with policy attributes / economic data.
\describe{
  \item{sk_emerytalna}{Variables related to the "skladka emerytalna" contribution.}
  \item{sk_rentowa}{Variables related to the "skladka rentowa" contribution.}
  \item{sk_chorobowa}{Variables related to the "skladka chorobowa" contribution}
  \item{sk_zdrowotna}{Variables related to the "skladka zdrowotna" contribution}
  \item{ppk}{Variables related to the "Pracownicze Plany Kapitalowe" contribution}
  \item{podatek}{Variables related to the tax contribution}
  \item{sl_plan2}{Variables related to the Student Loan Plan 2 contribution.}
  \item{sl_plan3}{Variables related to the Student Loan Plan 3 (postgraduate loan) contribution.}
  \item{earning_deciles}{Variables related to the national earning deciles.}
}
}
\usage{
pl_settings
}
\description{
Default PL Settings
}
\keyword{datasets}
